# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My the best API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/posts/create": {
            "post": {
                "tags": [
                    "post"
                ],
                "summary": "create post",
                "description": "create a new post",
                "operationId": "createPost",
                "requestBody": {
                    "description": "put fields related to a new post",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titel,description",
                                    "source"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "here is a title"
                                    },
                                    "decription": {
                                        "type": "string",
                                        "example": "this,id description "
                                    },
                                    "source": {
                                        "type": "string",
                                        "example": "sourcesource"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry,  Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Post": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "post"
                ],
                "summary": "Get post information",
                "description": "Returns post data",
                "operationId": "getPostById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/posts/?id={id}": {
            "delete": {
                "tags": [
                    "post"
                ],
                "summary": "Delete existing post",
                "description": "Deletes a record and returns no content",
                "operationId": "deletePost",
                "requestBody": {
                    "description": "put fields related to a  post to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/posts/update": {
            "post": {
                "tags": [
                    "post"
                ],
                "summary": "Update existing post",
                "description": "Returns updated post data",
                "operationId": "updatePost",
                "requestBody": {
                    "description": "put fields related to a  post to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id,titel,description",
                                    "source"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "here is an id"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "here is a title"
                                    },
                                    "decription": {
                                        "type": "string",
                                        "example": "this,id description "
                                    },
                                    "source": {
                                        "type": "string",
                                        "example": "sourcesource"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/getTopPosts": {
            "get": {
                "tags": [
                    "post"
                ],
                "summary": "Get posts",
                "description": "Returns sorted posts",
                "operationId": "getTopPosts",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "top posts",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "end date",
                        "schema": {
                            "type": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/getTopPostsSorted": {
            "get": {
                "tags": [
                    "post"
                ],
                "summary": "Get posts by interval of date day,month,week ...",
                "description": "Returns sorted posts by day , week,month",
                "operationId": "getTopPosts Sorted",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "0:day , 1:week , 2:month",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "value",
                        "in": "query",
                        "description": "0: current, 1:previous",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}